Add more retry in busybox when checking connectivity in network policy
This fix is to resolve an issue that the workload container in client Pod
on Windows might exit before it is attaching on the HNSEndpint. If the issue
happens, the DNS might not be able to resolve, and the related cases should
fail.

The fix inlcudes, 1) the connectivity check has 20 retry with an interval 2s
on Windows, 2) using "timeout -t" to control nc timeout as "nc -w" doesn't
work on Windows.